refactoring: moved all http APIs to services/api
