+++ src/store/peopleDB/sagas/removePerson.js
@@ -1,5 +1,5 @@
 import { call, put, takeEvery } from 'redux-saga/effects';
-import { deletePerson } from '../../../httpAPIs/reqresApi';
+import { deletePerson } from '../../../services/api/reqresApi';
 import { DELETE_PERSON_MIDDLEWARE } from '../../types';
 import { deletePerson as deletePersonAction } from '../actions';
